% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{download_cellage}
\alias{download_cellage}
\title{Download CellAge Data}
\usage{
download_cellage(outprefix = "cellage", outdir = "./")
}
\arguments{
\item{outprefix}{prefix for the name of output file, excluding the directory,
defaults to 'cellage'}

\item{outdir}{directory to save the file, defaults to the current directory}
}
\value{
returns the file path
}
\description{
Download CellAge Data
}
