% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugFunctions.R
\name{download_drug_interactions}
\alias{download_drug_interactions}
\title{Download drug gene interactions from DGIdb}
\usage{
download_drug_interactions(
  intfile = "https://dgidb.org/data/monthly_tsvs/2020-Nov/interactions.tsv"
)
}
\arguments{
\item{intfile}{link to the interactionfile from DGIdb}
}
\value{
returns a dataframe of 7 columns
}
\description{
Download drug gene interactions from DGIdb
}
