% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{download_drugage}
\alias{download_drugage}
\title{Download DrugAge Data}
\usage{
download_drugage(outprefix = "drugage", outdir = "./")
}
\arguments{
\item{outprefix}{prefix for the name of output file, excluding the directory,
defaults to 'drugage'}

\item{outdir}{directory to save the file, defaults to the current directory}
}
\value{
returns the file path
}
\description{
Download DrugAge Data
}
