% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{download_genage_human}
\alias{download_genage_human}
\title{Download GenAge Human Data}
\usage{
download_genage_human(outprefix = "genage_human", outdir = "./")
}
\arguments{
\item{outprefix}{prefix for the name of output file, excluding the directory,
defaults to 'genage_human'}

\item{outdir}{directory to save the file, defaults to the current directory}
}
\value{
returns the file path
}
\description{
Download GenAge Human Data
}
