% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{download_genage_model}
\alias{download_genage_model}
\title{Download GenAge Model Organism Data}
\usage{
download_genage_model(outprefix = "genage_models", outdir = "./")
}
\arguments{
\item{outprefix}{prefix for the name of output file, excluding the directory,
defaults to 'genage_models'}

\item{outdir}{directory to save the file, defaults to the current directory}
}
\value{
returns the file path
}
\description{
Download GenAge Model Organism Data
}
