% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{download_HAGR}
\alias{download_HAGR}
\title{Download GenAge, DrugAge, and CellAge Data}
\usage{
download_HAGR(outdir = "./")
}
\arguments{
\item{outdir}{directory to save the file, defaults to the current directory}
}
\value{
returns a data.frame with four columns including file paths to
GenAge Human (\code{human}), GenAge Model Organism (\code{model}), DrugAge (\code{drugage}),
and CellAge (\code{cellage}) files.
}
\description{
Download GenAge, DrugAge, and CellAge Data
}
