% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmapping.R
\name{get_all_geneIDs}
\alias{get_all_geneIDs}
\title{Get gene name, Ensembl IDs, Entrez IDs, Uniprot IDs, and gene biotype for all
genes}
\usage{
get_all_geneIDs(organism = "hsapiens")
}
\arguments{
\item{organism}{name of the organism: first letter of the genus immediately
followed by the specific name}
}
\value{
returns a list with three elements. First: \code{geneIDs}, a data frame
with 6 columns (Gene, EnsemblID, EntrezID, UniprotID, geneType, organism),
Second: \code{pkginfo}, a data frame with the package version information, Third:
\code{accessdate}: date of data accession (UTC time zone).
}
\description{
Get gene name, Ensembl IDs, Entrez IDs, Uniprot IDs, and gene biotype for all
genes
}
