% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoFuncs.R
\name{get_all_orthologs}
\alias{get_all_orthologs}
\title{Get all human orthologs for an organism}
\usage{
get_all_orthologs(organism = "mmusculus", target = "hsapiens")
}
\arguments{
\item{organism}{name of the organism: first letter of the genus immediately
followed by the specific name}
}
\value{
returns a list with three elements. First: \code{orthologs}, a data frame
with 9 columns (EnsemblID, humanEnsemblID, Type, humanInQuery, queryInHuman,
GOC, WGA, confidence, dataset), Second: \code{pkginfo}, a data frame with the
package version information, Third: \code{accessdate}: date of data accession
(UTC time zone).
}
\description{
Get all human orthologs for an organism
}
