% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmapping.R
\name{get_info4ChEMBLID}
\alias{get_info4ChEMBLID}
\title{Get molecule information for a ChEMBL ID}
\usage{
get_info4ChEMBLID(ChEMBLID)
}
\arguments{
\item{ChEMBLID}{ChEMBLID}
}
\value{
returns a data.frame with 19 columns: "ChEMBLID", "name", "type",
"therapeutic", "first_approval", "indication_class", "max_phase",
"atc_classifications", "black_box_warning", "oral", "topical", "natural",
"availability", "chirality", "inorganic", "smiles", "inchi", "inchi_key",
"synonyms"
}
\description{
Get molecule information for a ChEMBL ID
}
