% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugFunctions.R
\name{get_info4ChEMBLIDs}
\alias{get_info4ChEMBLIDs}
\title{Get molecule information for a vector of ChEMBL IDs}
\usage{
get_info4ChEMBLIDs(ChEMBLIDs)
}
\arguments{
\item{ChEMBLIDs}{a character vector of ChEMBL IDs}
}
\value{
returns a list with three elements. First: \code{ChEMBL_info}, a data
frame with 19 columns ( "ChEMBLID", "name", "type", "therapeutic",
"first_approval", "indication_class", "max_phase", "atc_classifications",
"black_box_warning", "oral", "topical", "natural", "availability",
"chirality", "inorganic", "smiles", "inchi", "inchi_key", "synonyms"),
Second: \code{pkginfo}, a data frame with the package version information,
Third: \code{accessdate}: date of data accession (UTC time zone).
}
\description{
Get molecule information for a vector of ChEMBL IDs
}
\examples{
get_info4ChEMBLIDs(c('CHEMBL413','CHEMBL1431'))

}
