% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmapping.R
\name{map_chemName2ChEMBLIDs}
\alias{map_chemName2ChEMBLIDs}
\title{Map chemicals to ChEMBL IDs}
\usage{
map_chemName2ChEMBLIDs(chem_names)
}
\arguments{
\item{chem_names}{a character vector including drug / chemical names}
}
\value{
returns a list with three elements. First: \code{ChEMBLIDs}, a data frame
with 3 columns (chemName, CID and ChEMBLID), Second: \code{pkginfo}, a data frame
with the package version information, Third: \code{accessdate}: date of data
accession (UTC time zone).
}
\description{
Map chemicals to ChEMBL IDs
}
\examples{
map_chemName2ChEMBLIDs(c("metformin", "rapamycin"))
}
