% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmapping.R
\name{map_chemName2CID}
\alias{map_chemName2CID}
\title{Chemical Name to PubChem CID}
\usage{
map_chemName2CID(chemname)
}
\arguments{
\item{chemname}{Chemical name}
}
\value{
returns pubchem CID
}
\description{
Chemical Name to PubChem CID
}
