% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmapping.R
\name{map_chemName2CIDs}
\alias{map_chemName2CIDs}
\title{Map chemicals to PubChem CIDs}
\usage{
map_chemName2CIDs(chem_names)
}
\arguments{
\item{chem_names}{a character vector including names of the drugs / chemicals}
}
\value{
returns a list with three elements. First: \code{CIDs}, a data frame with
2 columns (chemName and CID), Second: \code{pkginfo}, a data frame with the
package version information, Third: \code{accessdate}: date of data accession
(UTC time zone).
}
\description{
Map chemicals to PubChem CIDs
}
\examples{
map_chemName2CIDs(c("metformin", "rapamycin"))
}
