% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmapping.R
\name{map_CID2ChEMBLID}
\alias{map_CID2ChEMBLID}
\title{PubChem CID to ChEMBL ID}
\usage{
map_CID2ChEMBLID(cid)
}
\arguments{
\item{cid}{PubChem CID}
}
\value{
returns ChEMBLID
}
\description{
PubChem CID to ChEMBL ID
}
