% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmapping.R
\name{map_CID2ChEMBLIDs}
\alias{map_CID2ChEMBLIDs}
\title{Map PubChem CIDs to ChEMBL IDs}
\usage{
map_CID2ChEMBLIDs(cids)
}
\arguments{
\item{cids}{a character vector including PubChem CIDs}
}
\value{
returns a list with three elements. First: \code{ChEMBLIDs}, a data frame
with 2 columns (CID and ChEMBLID), Second: \code{pkginfo}, a data frame with the
package version information, Third: \code{accessdate}: date of data accession
(UTC time zone).
}
\description{
Map PubChem CIDs to ChEMBL IDs
}
\examples{
map_CID2ChEMBLIDs(c("4091", "5284616"))
}
