% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{tidy_cellage}
\alias{tidy_cellage}
\title{Tidy CellAge Data}
\usage{
tidy_cellage(path2file)
}
\arguments{
\item{path2file}{path to cellage csv file}
}
\value{
returns a data frame with 7 columns: organism, dataset, Gene,
cancer_type, senescence_effect, description, notes
}
\description{
Tidy CellAge Data
}
