% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{tidy_drugage}
\alias{tidy_drugage}
\title{Tidy DrugAge Data}
\usage{
tidy_drugage(path2file)
}
\arguments{
\item{path2file}{path to drugage csv file}
}
\value{
returns a data frame with 10 columns: 'organism','dataset','compound_name',
'strain','dosage','avg_lifespan_change','max_lifespan_change','gender','significance','pubmed_id'
}
\description{
Tidy DrugAge Data
}
