% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{tidy_genage_human}
\alias{tidy_genage_human}
\title{Tidy GenAge Human Data}
\usage{
tidy_genage_human(path2file)
}
\arguments{
\item{path2file}{path to genage human csv file}
}
\value{
returns a data frame with two columns: Gene and Why
}
\description{
Tidy GenAge Human Data
}
