% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRdata.R
\name{tidy_genage_model}
\alias{tidy_genage_model}
\title{Tidy GenAge Model Organism Data}
\usage{
tidy_genage_model(path2file)
}
\arguments{
\item{path2file}{path to genage model csv file}
}
\value{
returns a data frame with 6 columns: 'Gene','avgChange','Direction',
'Type','organism','dataset'
}
\description{
Tidy GenAge Model Organism Data
}
