% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoFuncs.R
\name{tidy_orthologs}
\alias{tidy_orthologs}
\title{Get all IDs and human orthologs for an organism}
\usage{
tidy_orthologs(organism = "mmusculus", target = "hsapiens")
}
\arguments{
\item{organism}{name of the organism: first letter of the genus immediately
followed by the specific name}
}
\value{
returns a list with three elements. First: \code{orthologs}, a data frame
with 13 columns (Gene, EnsemblID, EntrezID, UniprotID, geneType, dataset,
humanEnsemblID, Type, humanInQuery, queryInHuman, GOC, WGA, confidence),
Second: \code{pkginfo}, a data frame with the package version information,
Third: \code{accessdate}: date of data accession (UTC time zone).
}
\description{
Get all IDs and human orthologs for an organism
}
